fix(ui): keep config search off startup constraints
