+++ ui/src/components/config-form.constraints.ts
@@ -4,6 +4,7 @@ import {
   jsonSchemaValuesEqual,
 } from "@openclaw/normalization-core/json-schema";
 import { asFiniteNumber as finiteNumber } from "@openclaw/normalization-core/number-coercion";
+import { arrayItemSchema, collectAllOfSchemas, combinedSchema } from "./config-form.array-items.ts";
 import { decimalRational } from "./config-form.numeric.ts";
 import { schemaType, type JsonSchema } from "./config-form.shared.ts";
 
@@ -109,27 +110,6 @@ type EffectiveNumericBound = {
   value?: number;
   exclusive: boolean;
 };
-
-function collectAllOfSchemas(schema: JsonSchema): JsonSchema[] {
-  const result: JsonSchema[] = [];
-  const pending = [schema];
-  const seen = new Set<JsonSchema>();
-  while (pending.length > 0) {
-    const current = pending.pop();
-    if (!current || seen.has(current)) {
-      continue;
-    }
-    seen.add(current);
-    result.push(current);
-    for (let index = (current.allOf?.length ?? 0) - 1; index >= 0; index -= 1) {
-      const entry = current.allOf?.[index];
-      if (entry) {
-        pending.push(entry);
-      }
-    }
-  }
-  return result;
-}
 
 function effectiveNumericBound(
   schemas: JsonSchema[],
@@ -239,16 +219,6 @@ export function objectPropertyKeys(schema: JsonSchema): string[] {
   );
 }
 
-function combinedSchema(candidates: JsonSchema[]): JsonSchema | undefined {
-  const base = candidates.find((candidate) => schemaType(candidate) !== undefined) ?? candidates[0];
-  return !base || candidates.length === 1
-    ? base
-    : {
-        ...base,
-        allOf: [...(base.allOf ?? []), ...candidates.filter((candidate) => candidate !== base)],
-      };
-}
-
 export function objectAdditionalPropertiesSchema(
   schema: JsonSchema,
 ): JsonSchema | false | undefined {
@@ -319,39 +289,6 @@ export function objectPropertySchema(schema: JsonSchema, key: string): JsonSchem
       .map((entry) => ownPropertySchema(entry, key))
       .filter((property): property is JsonSchema => property !== undefined),
   );
-}
-
-export function arrayItemSchema(schema: JsonSchema, index: number): JsonSchema | undefined {
-  const candidates: JsonSchema[] = [];
-  for (const entry of collectAllOfSchemas(schema)) {
-    if (Array.isArray(entry.items)) {
-      const item =
-        entry.items[index] ??
-        (entry.additionalItems && typeof entry.additionalItems === "object"
-          ? entry.additionalItems
-          : undefined);
-      if (item) {
-        candidates.push(item);
-      }
-    } else if (entry.items) {
-      candidates.push(entry.items);
-    }
-  }
-  return combinedSchema(candidates);
-}
-
-export function arrayItemSchemaIndexes(schema: JsonSchema): number[] {
-  let count = 0;
-  for (const entry of collectAllOfSchemas(schema)) {
-    if (Array.isArray(entry.items)) {
-      const hasTypedTail =
-        entry.additionalItems !== null && typeof entry.additionalItems === "object";
-      count = Math.max(count, entry.items.length + (hasTypedTail ? 1 : 0));
-    } else if (entry.items) {
-      count = Math.max(count, 1);
-    }
-  }
-  return Array.from({ length: count }, (_, index) => index);
 }
 
 export function arrayConstraintCandidates(
