+++ ui/src/components/config-form.search.ts
@@ -1,6 +1,6 @@
 import type { ConfigUiHints } from "../api/types.ts";
 import { normalizeLowercaseStringOrEmpty } from "../lib/string-coerce.ts";
-import { arrayItemSchema, arrayItemSchemaIndexes } from "./config-form.constraints.ts";
+import { arrayItemSchema, arrayItemSchemaIndexes } from "./config-form.array-items.ts";
 import { hintForPath, humanize, schemaType, type JsonSchema } from "./config-form.shared.ts";
 
 export type ConfigSearchCriteria = {
